requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
nltk
wordcloud
python-docx
PyPDF2
